+++ src/compile/header/index.ts
@@ -4,7 +4,6 @@
 import {Axis as VgAxis, AxisOrient, TitleConfig as VgTitleConfig} from 'vega';
 import {isArray} from 'vega-util';
 import {Config} from '../../config';
-import {FacetFieldDef} from '../../facet';
 import {vgField} from '../../fielddef';
 import {
   HEADER_LABEL_PROPERTIES,
@@ -14,6 +13,7 @@ import {
   HeaderConfig
 } from '../../header';
 import {isSortField} from '../../sort';
+import {FacetFieldDef} from '../../spec/facet';
 import {keys} from '../../util';
 import {VgComparator, VgMarkGroup} from '../../vega.schema';
 import {formatSignalRef} from '../common';
